requirements.txt
fastapi
uvicorn
transformers
torch
streamlit
sentence-transformers
faiss-cpu
requests
pydantic
